requirements.txt
asttokens==3.0.0
certifi==2025.8.3
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
distlib==0.4.0
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.19.1
fonttools==4.59.0
greenlet==3.2.3
ipykernel==6.29.5
ipython==8.18.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.5
packaging==24.2
pandas==2.3.1
parso==0.8.4
Pillow==9.5.0
pipenv==2025.0.4
platformdirs==4.3.7
polars-lts-cpu==1.33.0
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
pyzmq==26.4.0
schedule==1.2.2
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.41
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
virtualenv==20.34.0
wcwidth==0.2.13
wordcloud==1.9.4
